TW-230: refactor old widget and add mixin
