+++ lib/pages/new_private_chat/new_private_chat.dart
@@ -7,10 +7,9 @@ import 'package:fluffychat/mixin/comparable_presentation_contact_mixin.dart';
 import 'package:fluffychat/pages/new_private_chat/fetch_contacts_controller.dart';
 import 'package:fluffychat/pages/new_private_chat/new_private_chat_view.dart';
 import 'package:fluffychat/pages/new_private_chat/search_contacts_controller.dart';
-import 'package:fluffychat/presentation/model/presentation_contact.dart';
-import 'package:fluffychat/widgets/matrix.dart';
+import 'package:fluffychat/presentation/mixin/go_to_direct_chat_mixin.dart';
 import 'package:flutter/material.dart';
-import 'package:future_loading_dialog/future_loading_dialog.dart';
+
 import 'package:matrix/matrix.dart';
 import 'package:vrouter/vrouter.dart';
 
@@ -22,7 +21,7 @@ class NewPrivateChat extends StatefulWidget {
 }
 
 class NewPrivateChatController extends State<NewPrivateChat> 
-  with ComparablePresentationContactMixin {
+  with ComparablePresentationContactMixin, GoToDirectChatMixin {
 
   final searchContactsController = SearchContactsController();
   final fetchContactsController = FetchContactsController();
@@ -69,18 +68,6 @@ class NewPrivateChatController extends State<NewPrivateChat>
   void toggleContactsList() {
     isShowContactsNotifier.value = !isShowContactsNotifier.value;
     fetchContactsController.haveMoreCountactsNotifier.value = isShowContactsNotifier.value;
-  }
-
-  void goToChatScreen({required PresentationContact contact}) {
-    showFutureLoadingDialog(
-      context: context,
-      future: () async {
-        if (contact.matrixId != null && contact.matrixId!.isNotEmpty) {
-          final roomId = await Matrix.of(context).client.startDirectChat(contact.matrixId!);
-          VRouter.of(context).toSegments(['rooms', roomId]);
-        }
-      },
-    );
   }
 
   void goToNewGroupChat() {
